Code from topcon and student project upmerge

+++ core/alignment/radialFunc.cpp
@@ -38,7 +38,7 @@ void ModelToRadialCorrection::getModel(const RadialCorrection &correction, doubl
         in[count++] = correction.mParams.p2;
     }
     for (int i = 0; i < mPolynomPower; i++) {
-        if (i < correction.mParams.koeff.size()) {
+        if (i < (int)correction.mParams.koeff.size()) {
             in[count++] = correction.mParams.koeff[i];
         } else {
             in[count++] = 0.0;
